requirements.txt
python-telegram-bot==20.7
telethon==1.34.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2024.1
